Prepare code for mutation testing
This commit adds the optional with_mutagen feature, which brings in the in-development version of mutagen. It adds the attribute to all the DNS parsers, opting them in to mutation testing.

The mutagen tool itself recommends using the Git master version; however, it currently lacks the --package option which I added in my branch.