Make incorrect label lengths a hard protocol error
This commit makes all record types properly check the length of the data after they have read it, returning an error if there is a mismatch. Now, no record types ignore their length argument. Some of them used to log a warning, but now they all do, as well as returning the error.

Doing this required changing the read_labels function so that it returns the number of bytes read without recursing, as this number is not necessarily the same as the number of bytes in the string plus one (for the null terminator).

For clarity, the WrongLength error case has been split in two, to tell the difference between where the data disagrees with the DNS protocol, and where the data disagrees with the length field in the packet.