Limit code to supporting one query per request
The dns crate technically allowed more than one query per request, as that's what's supported by the DNS spec, but in practice, only one is sent, and there are no plans to add an option to send more. So the singleton vector has been inlined.

Doing this also removed a cast from usize to u16.