Be careful when sending too much OPT data
This is another integer conversion fix. Even though no data is sent in OPT packets in requests right now, there was a conversion from usize to 16 as it wrote the size of the data. The best thing to do would be to change the design so that request!OPT and response!OPT don't share the same type (with the request kind not even having the data field to send), but at least with this, the intent is made clear.