Introduce Labels type to hold labels
There was a bug in the function to encode labels: the length of a label segment must fit in a u8, so there's a maximum length of 255; however, the domain comes from the user when making requests, and the length was lossily casted to u8 (integer conversion again). This meant that the packet would effectively contain junk data after the domain, which DNS servers did not like. This led to a further crash in the error-handling code when a server sends back an empty response (this is not fixed yet).

The simplest way to deal with this would be to use u8::try_from and unwrap when writing the label to the buffer, but this isn't very user-friendly, as the error should be detected up-front. So the Labels type was introduced, which validates the length of its input's segments. Everything that encodes or decodes a domain name now uses this type.