Tag tests and use Cargo features for quick mode
While working on the DNS wire format crate, I've found it annoying that my standard compile-test-xtests loop involves linking in the TLS and HTTPS dependencies, then running the tests for every transport, when I know all I've changed is the protocol parsing parts.

So now there's the "all-quick" Justfile target, which disables these features and runs fewer of the tests. This has noticeable build speed improvements:

• building and linking from a modified 'dns' crate has gone from 3.5s → 1.6s
• a clean build has gone from 28s → 14s
• an entire clean cycle has gone from 38s → 17s

The "all" Justfile target remains the default one, as it's the recommended way to test dog with all its features. But the increased linking speed will be helpful, development-wise.