Replace Tokio with std
This commit removes a load of dependencies and replaces them with code that just uses std. dog was only dealing with one request at a time anyway, so all the async stuff was just unnecessary overhead.

• tokio for UDP and TCP transports gets replaced with std::net
• native-tls + tokio-tls get replaced with just native-tls
• hyper and hyper-tls also get replaced with native-tls and httparse, where we write the HTTP headers ourselves
• async-trait is no longer necessary since the trait is sync now

Now, all the four main transport types work the same way — create a buffer and fill it with bytes read from a stream (or datagram for UDP). This commit also removes like 40 dependencies, and just about halves compilation time.