Handle truncated responses and standardise netcode
This commit adds a new network error variant: TruncatedResponse, which gets returned when a read call yields 0 bytes. This used to be an outright panic, but it was possible to trigger this by pointing dog to a non-DNS TCP server that doesn't send anything back when you give it a DNS request.

The TLS transport is meant to perform the same length-prefix wrangling as the TCP transport, and sure enough, an encrypted non-DNS TCP server resulted in the same thing. Some functions have been extracted to allow for code sharing. This resulted in the buffer being put on the heap, rather than on the stack, so we can return it.

Some more general work was also done to make the four main transport types more consistent with each other.