Add check when using --https without a nameserver
Here, we add an explicit check when the user passes the --https argument without a nameserver. What currently happens is that dog tries to use the system default resolver, which won't be a URL when it should be.

This was raised in GH-5. The panic is still there, and it's still simple to trigger, so it's not 100% fixed yet, but at least it's now harder to trigger _accidentally_.