Allow multiple messages in TXT record parsing
This commit fixes a bug in the TXT record parser, where it assumed it could only hold one message. According to the standard, multiple messages can be present in a single TXT record, so now we continually read messages in a loop until the size limit is hit.

Fixes GH-18.