Move huge unit tests out of wire module
I figured out why cargo-mutagen wasn't picking up any coverage for the integration tests: it was only being run when the 'test' flag was enabled, so it only picked up unit tests.

This was the only reason that these long tests were unit tests, cluttering up the wire.rs file, so now that's changed, they've been moved to integration test files. We also now log a warning if the feature somehow gets enabled when building the main application, which should offset the danger of this feature technically being available to non-test builds.