Resolver-related code cleanup
This commit introduces a new exit code (4) for when no system resolver can be found, instead of just crashing.

It also introduces ResolverType, which is similar to code that used to be present. While the previous code passed all its unit tests, there was system-level code (obtaining the resolvers) being called from those unit tests, which isn't something unit tests are meant to be doing — there's no guarantee that they'll continue to work on systems with different network conifgurations. Because ResolverType is just a type, it can be used in tests. (Resolver has had its PartialEq impl stripped, and there's now a warning, to stop this from happening again.)

The refactor lets us remove some Clippy allowances, which is always nice.