Handle single-word domain free arguments
This commit fixes a regression caused by the fix for GH-14, which made _any_ single-word argument be treated as though it should be a DNS class or type name. So `dog aoeu` would attempt to parse `aoeu` as a type, fail, and display an error. Now, it correctly does a domain lookup for `aoeu`.

The root cause was me trying to use the same functions for both parsing free arguments and named arguments, when they do two different things.